The initial aurora in "FoxtTail_Experiment" Scene
Color-height functionality is applied, looking further options to make the lights stay a little longer and also figuring out how to remove the gravity effect, even though the gravity modifier is set to 0, the effect remains, suspecting it to be coming from setting the simulation space to World. However, that's necessary to achieve the trail effect!
Will  take some more tile